+++ client/src/components/SideNavPages/Owner/ExpenseList.jsx
@@ -46,6 +46,8 @@ const ExpenseList = () => {
 
   // DELETE EXPENSE API CALL
   const Delete = async (expense_id) => {
+    const confirmDelete = window.confirm("Are you sure?");
+    if (!confirmDelete) return;
     try {
       await axios.delete(
         `${import.meta.env.VITE_API_URL}/api/va/auth/expense-delete`,
